ensure user is invite creator before deleting

+++ lib/collaborlist_web/live/invite_modal.ex
@@ -121,18 +121,23 @@ defmodule CollaborlistWeb.Live.InviteModal do
 
   def handle_event("delete", %{"invite_code" => invite_code}, socket) do
     invite = Invites.get_invite!(invite_code)
-    {:ok, _invite} = Invites.delete_invite(invite)
 
-    invites = socket.assigns.invites
+    if Invites.invite_creator?(socket.assigns.current_user, invite.invite_code) do
+      {:ok, _invite} = Invites.delete_invite(invite)
 
-    invites_after_delete =
-      invites
-      |> List.delete_at(Enum.find_index(invites, fn i -> i.invite_code == invite_code end))
+      invites = socket.assigns.invites
 
-    {:noreply,
-     assign(socket,
-       invites: invites_after_delete
-     )}
+      invites_after_delete =
+        invites
+        |> List.delete_at(Enum.find_index(invites, fn i -> i.invite_code == invite_code end))
+
+      {:noreply,
+       assign(socket,
+         invites: invites_after_delete
+       )}
+    else
+      {:noreply, assign(socket, error_message: "you can't do that")}
+    end
   end
 
   def handle_event(
